Some minor improvements
In the `LimitController` I replaced the response handling with a method
from Symfony. No need to handle it manually with twig

The `CustomerRepository` did not implement its interface. And then I
moved the documentation to the interface so that future implemenations
can reference it as well with the `@inheritdoc` syntax.